[backend] Separate ACP and UCP into protocol namespaces
