--- src/merchant/api/routes/ucp/__init__.py
@@ -1,1 +1,0 @@
-"""UCP API routes package."""
