+++ src/merchant/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain layer for protocol-agnostic merchant business logic."""
