+++ src/merchant/domain/checkout/__init__.py
@@ -1,0 +1,1 @@
+"""Protocol-agnostic checkout domain models and services."""
