+++ src/merchant/protocols/__init__.py
@@ -1,0 +1,1 @@
+"""Protocol-specific adapters for merchant APIs."""
