+++ src/merchant/protocols/acp/__init__.py
@@ -1,0 +1,1 @@
+"""ACP protocol adapter package."""
