+++ src/merchant/protocols/acp/api/__init__.py
@@ -1,0 +1,1 @@
+"""ACP API layer."""
