+++ src/merchant/protocols/acp/api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""ACP route handlers."""
