+++ src/merchant/protocols/acp/api/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""ACP wire schemas."""
