+++ src/merchant/protocols/acp/api/schemas/checkout.py
@@ -1,0 +1,6 @@
+"""ACP checkout wire schemas.
+
+The ACP protocol uses the checkout domain model set directly as its wire shape.
+"""
+
+from src.merchant.domain.checkout.models import *  # noqa: F401,F403
