+++ src/merchant/protocols/ucp/__init__.py
@@ -1,0 +1,1 @@
+"""UCP protocol adapter package."""
