+++ src/merchant/protocols/ucp/api/__init__.py
@@ -1,0 +1,1 @@
+"""UCP API layer."""
