+++ src/merchant/protocols/ucp/api/routes/__init__.py
@@ -1,0 +1,1 @@
+"""UCP route handlers."""
