+++ src/merchant/protocols/ucp/api/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""UCP wire schemas."""
