+++ src/merchant/protocols/ucp/services/__init__.py
@@ -1,0 +1,1 @@
+"""UCP protocol services."""
