+++ src/merchant/services/__init__.py
@@ -1,1 +1,1 @@
-# Business logic services
+# Shared business services (protocol-neutral)
