+++ tests/merchant/architecture/__init__.py
@@ -1,0 +1,1 @@
+"""Architecture-level backend tests."""
